Monifier.BusinessLogic.Queries/Distribution/DistributionQueries.cs
using System.Threading.Tasks;
using Monifier.BusinessLogic.Contract.Auth;
using Monifier.BusinessLogic.Contract.Distribution;
using Monifier.BusinessLogic.Distribution;
using Monifier.DataAccess.Contract;

namespace Monifier.BusinessLogic.Queries.Distribution
{
    public class DistributionQueries : IDistributionQueries
    {
        private readonly IUnitOfWork _unitOfWork;
        private readonly ICurrentSession _currentSession;

        public DistributionQueries(IUnitOfWork unitOfWork, ICurrentSession currentSession)
        {
            _unitOfWork = unitOfWork;
            _currentSession = currentSession;
        }
        
        public async Task<DistributionBoard> GetDistributionBoard()
        {
            var board = new DistributionBoard();
            await board.Load(_unitOfWork, _currentSession.UserId);
            return board;
        }
    }
}